requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
event-emitter-js==0.1.2
executing==2.2.0
fastapi==0.115.8
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
logfire==3.5.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
node-semver==0.9.0
oauthlib==3.2.2
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pillow==11.1.0
playwright==1.50.0
playwright-stealth==1.0.6
PlaywrightSafeThread==0.5.5
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
segno==1.6.1
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
WPP_Whatsapp @ git+https://github.com/3mora2/WPP_Whatsapp@e45161086704afc369e8d550105e1f598216b1eb
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
